Fix file descriptor leak in the plugin for episode generation (#1339)
Before this changes in this commit, the plugin opens an OutputStream to
the episode file and flushes it after the document has successfully been
constructed. However, it doesn't close the stream explicitly and instead
relies on the GC collecting the stream object, which in turn closes it
through the finalize method.

Until the GC decides to collect the object (which it may chose to never
do), the plugin will hold a file descriptor open. On Windows, this locks
the file and prevents any further actions on it, like rewriting it by
re-running XJC.

This is not a problem when XJC is run in a short-lived JVM process, such
as when run from the command line or through the Maven plugin. But on
long-lived processes, such as through a Gradle daemon, it will lock the
file and prevent it to run again.

This commit ensures that the OutputStream is always closed at the end of
the run method of the plugin, even if an exception is thrown.

Signed-off-by: Bjørn Vester <bmv@netcompany.com>