feat(xjc): call postProcessModel with DTD models
When code is generated from DTD schema, the hook function
`postProcessModel` was not called. The code to do so was moved
from global utility class to `BGMBuilder` class in the old code more
than five years ago with the effect, that the hook was not called
anymore for any other input than XSD.

A Junit 4 test is included.

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>